+++ src/main/java/wooteco/subway/auth/dto/LoginMember.java
@@ -1,0 +1,13 @@
+package wooteco.subway.auth.dto;
+
+public class LoginMember {
+    private Long id;
+
+    public LoginMember(Long id) {
+        this.id = id;
+    }
+
+    public Long getId() {
+        return id;
+    }
+}
